requirements.txt
opencv-python
opencv-contrib-python
mediapipe
numpy
pandas
joblib
lightgbm
scikit-learn
twilio
pygame
python-dotenv
features
detector